+++ x/posts/internal/types/keys.go
@@ -5,14 +5,20 @@ const (
 	RouterKey  = ModuleName
 	StoreKey   = ModuleName
 
-	PostStorePrefix         = "post:"
-	LastPostIDStoreKey      = "last_post_id"
-	PostCommentsStorePrefix = "comments:"
+	PostStorePrefix                 = "post:"
+	LastPostIDStoreKey              = "last_post_id"
+	PostCommentsStorePrefix         = "comments:"
+	PostReactionsStorePrefix        = "reactions:"
+	MaxOptionalDataFieldsNumber     = 10
+	MaxOptionalDataFieldValueLength = 200
 
-	LikesStorePrefix = "likes:"
+	ActionCreatePost         = "create_post"
+	ActionEditPost           = "edit_post"
+	ActionAddPostReaction    = "add_post_reaction"
+	ActionRemovePostReaction = "remove_post_reaction"
 
-	ActionCreatePost = "create_post"
-	ActionEditPost   = "edit_post"
-	ActionLikePost   = "like_post"
-	ActionUnlikePost = "unlike_post"
+	// Queries
+	QuerierRoute = ModuleName
+	QueryPost    = "post"
+	QueryPosts   = "posts"
 )